[Structure] refactored Mongoose code
